removed unneeded Dashboard component
