--- client/components/Host/HostDashboard.js
@@ -1,19 +1,0 @@
-import React from 'react'
-import Dashboard from './Dashboard'
-
-class HostDashboard extends React.Component {
-  constructor(props) {
-    super(props)
-
-    this.state = {
-      user: {}
-    }
-  }
-
-  render() {
-    //check for user authentication and redirect here
-    return <Dashboard />
-  }
-}
-
-export default HostDashboard
