+++ client/components/index.js
@@ -11,4 +11,4 @@ export {default as GuestLogin} from './GuestLogin'
 export {default as PartyRoom} from './PartyRoom'
 export {default as Chat} from './Chat'
 export {default as Room} from './PartyRoom'
-export {default as HostDashboard} from './Host/HostDashboard'
+export {default as HostDashboard} from './HostDashboard'
